#5.1 Hash User Password - Mã Hóa Mật Khẩu với Bcrypt | SERIES FULLSTACK - JWT, Node.JS & React
