+++ HoiDanIT/fullstack web/jwtBackend/src/controller/homeController.js
@@ -1,11 +1,4 @@
-import mysql from 'mysql2'
-
-// create the connection to database
-const connection = mysql.createConnection({
-  host: 'localhost',
-  user: 'root',
-  database: 'jwt'
-})
+import userService from '../service/userService'
 
 const handleHelloWord = (req, res) => {
   return res.render('home.ejs')
@@ -13,7 +6,6 @@ const handleHelloWord = (req, res) => {
 
 const handleUserPage = (req, res) => {
   // model => get data from database
-
   return res.render('user.ejs')
 }
 
@@ -22,15 +14,8 @@ const handleCreateNewUser = (req, res) => {
   let password = req.body.password
   let username = req.body.username
 
-  connection.query(
-    'INSERT INTO users (email, password, username) VALUES (?, ?, ?)',
-    [email, password, username],
-    function (err, results, fields) {
-      if (err) {
-        console.log(err)
-      }
-    }
-  )
+  // userService.createNewUser(email, password, username)
+  userService.getUserList()
 
   return res.send('handleCreateNewUser')
 }
