+++ HoiDanIT/fullstack web/jwtBackend/src/views/user.ejs
@@ -15,7 +15,10 @@
       <form method="POST" action="/users/create-user">
         <div class="container">
           <div class="row">
-            <div class="mb-3 mt-3 col-6">
+            <div class="pt-3">
+              <h3>Create a new user:</h3>
+            </div>
+            <div class="mb-3 col-6">
               <label for="email" class="form-label">Email address</label>
               <input type="email" class="form-control" id="email" name="email" required />
             </div>
